move cleverF function to cleverF.ts
