+++ src/actions/include-all.ts
@@ -17,6 +17,7 @@ import './commands/scroll';
 
 // plugin
 import './plugins/camelCaseMotion';
+import './plugins/cleverF';
 import './plugins/easymotion/easymotion.cmd';
 import './plugins/easymotion/registerMoveActions';
 import './plugins/sneak';
