+++ src/actions/include-plugins.ts
@@ -1,5 +1,6 @@
 // plugin
 import './plugins/camelCaseMotion';
+import './plugins/cleverF';
 import './plugins/easymotion/easymotion.cmd';
 import './plugins/easymotion/registerMoveActions';
 import './plugins/sneak';
