+++ src/actions/motion.ts
@@ -839,7 +839,7 @@ class MoveToColumn extends BaseMovement {
  * @param count number of times to look
  * @param direction direction to look in
  */
-function findHelper(
+export function findHelper(
   vimState: VimState,
   start: Position,
   char: string,
@@ -865,8 +865,6 @@ function findHelper(
   return undefined;
 }
 
-let posF = [-1, -1];
-let beforeCleverFAction = '';
 @RegisterAction
 class MoveFindForward extends BaseMovement {
   keys = ['f', '<character>'];
@@ -908,39 +906,6 @@ class MoveFindForward extends BaseMovement {
 }
 
 @RegisterAction
-class ActionCleverForwardCommand extends BaseMovement {
-  override modes = [Mode.Normal];
-  keys = ['f'];
-
-  public override doesActionApply(vimState: VimState, keysPressed: string[]): boolean {
-    return configuration.cleverF && super.doesActionApply(vimState, keysPressed);
-  }
-
-  public override async execActionWithCount(
-    position: Position,
-    vimState: VimState,
-    count: number
-  ): Promise<Position | IMovement> {
-    count ||= 1;
-    if (arrayEqual(posF, [position.line, position.character]) && beforeCleverFAction === 'f') {
-      const action = new MoveRepeat([';'], true);
-      return action.execActionWithCount(position, vimState, count);
-    } else if (
-      arrayEqual(posF, [position.line, position.character]) &&
-      beforeCleverFAction === 'F'
-    ) {
-      const action = new MoveRepeatReversed([','], true);
-      beforeCleverFAction = 'f';
-      return action.execActionWithCount(position, vimState, count);
-    } else {
-      await vimState.setCurrentMode(Mode.CleverFForwardMode);
-      beforeCleverFAction = 'f';
-      return position;
-    }
-  }
-}
-
-@RegisterAction
 class MoveFindBackward extends BaseMovement {
   keys = ['F', '<character>'];
 
@@ -968,107 +933,6 @@ class MoveFindBackward extends BaseMovement {
       return failedMovement(vimState);
     }
 
-    return result;
-  }
-}
-
-@RegisterAction
-class ActionCleverFBackwardCommand extends BaseMovement {
-  override modes = [Mode.Normal];
-  keys = ['F'];
-
-  public override doesActionApply(vimState: VimState, keysPressed: string[]): boolean {
-    return configuration.cleverF && super.doesActionApply(vimState, keysPressed);
-  }
-
-  public override async execActionWithCount(
-    position: Position,
-    vimState: VimState,
-    count: number
-  ): Promise<Position | IMovement> {
-    count ||= 1;
-    try {
-      if (arrayEqual(posF, [position.line, position.character]) && beforeCleverFAction === 'F') {
-        const action = new MoveRepeat([';'], true);
-        return action.execActionWithCount(position, vimState, count);
-      } else if (
-        arrayEqual(posF, [position.line, position.character]) &&
-        beforeCleverFAction === 'f'
-      ) {
-        const action = new MoveRepeatReversed([','], true);
-        beforeCleverFAction = 'F';
-        return action.execActionWithCount(position, vimState, count);
-      } else {
-        await vimState.setCurrentMode(Mode.CleverFBackwardMode);
-        beforeCleverFAction = 'F';
-        return position;
-      }
-    } finally {
-      console.log('test');
-    }
-  }
-}
-
-const arrayEqual = (a: number[], b: number[]): boolean => {
-  if (!Array.isArray(a)) return false;
-  if (!Array.isArray(b)) return false;
-  if (a.length !== b.length) return false;
-  for (let i = 0, n = a.length; i < n; ++i) {
-    if (a[i] !== b[i]) return false;
-  }
-  return true;
-};
-
-@RegisterAction
-class MoveCleverFFindForward extends BaseMovement {
-  override modes = [Mode.CleverFForwardMode];
-  keys = ['<character>'];
-  public override async execActionWithCount(
-    position: vscode.Position,
-    vimState: VimState,
-    count: number
-  ): Promise<vscode.Position | IMovement> {
-    await vimState.setCurrentMode(Mode.Normal);
-    posF = [position.line, position.character];
-    count ||= 1;
-    const toFind = Notation.ToControlCharacter(this.keysPressed[0]);
-    let result = findHelper(vimState, position, toFind, count, 'forward');
-    vimState.lastSemicolonRepeatableMovement = new MoveCleverFFindForward(this.keysPressed, true);
-    vimState.lastCommaRepeatableMovement = new MoveCleverFFindBackward(this.keysPressed, true);
-    if (!result) {
-      return failedMovement(vimState);
-    }
-    if (vimState.recordedState.operator) {
-      result = result.getRight();
-    }
-    posF = [result.line, result.character];
-    return result;
-  }
-}
-
-@RegisterAction
-class MoveCleverFFindBackward extends BaseMovement {
-  override modes = [Mode.CleverFBackwardMode];
-  keys = ['<character>'];
-  public override async execActionWithCount(
-    position: vscode.Position,
-    vimState: VimState,
-    count: number
-  ): Promise<vscode.Position | IMovement> {
-    await vimState.setCurrentMode(Mode.Normal);
-    posF = [position.line, position.character];
-    count ||= 1;
-    const toFind = Notation.ToControlCharacter(this.keysPressed[0]);
-    let result = findHelper(vimState, position, toFind, count, 'backward');
-    vimState.lastSemicolonRepeatableMovement = new MoveCleverFFindBackward(this.keysPressed, true);
-    vimState.lastCommaRepeatableMovement = new MoveCleverFFindForward(this.keysPressed, true);
-    if (!result) {
-      return failedMovement(vimState);
-    }
-    if (vimState.recordedState.operator) {
-      result = result.getRight();
-    }
-    posF = [result.line, result.character];
     return result;
   }
 }
@@ -1147,7 +1011,7 @@ class MoveTilBackward extends BaseMovement {
 }
 
 @RegisterAction
-class MoveRepeat extends BaseMovement {
+export class MoveRepeat extends BaseMovement {
   keys = [';'];
 
   public override async execActionWithCount(
@@ -1164,7 +1028,7 @@ class MoveRepeat extends BaseMovement {
 }
 
 @RegisterAction
-class MoveRepeatReversed extends BaseMovement {
+export class MoveRepeatReversed extends BaseMovement {
   keys = [','];
 
   public override async execActionWithCount(
